Add PlacemarkDescribable protocol. Move CLPlacemark computed properties to protocol.  Conform CLPlacemark and Placemark to PlacemarkDescribable
